+++ src/utils/logger.tsx
@@ -1,12 +1,6 @@
 import { h } from '@stencil/core';
 import { report } from './errorReport';
 
-interface ErrorDetail {
-  componentName?: string;
-  error: string;
-  npmVersion: string;
-}
-
 interface StencilComponent {
   constructor: {
     name: string;
@@ -64,12 +58,11 @@ export default function logger<T>() {
         }
         return rendered;
       } catch (e) {
-        const detail: ErrorDetail = {
+        report({
+          code: e.message,
           componentName: target.constructor.name,
-          error: e.message,
-          npmVersion: '<@NPM_PACKAGE_VERSION@>',
-        };
-        report(detail);
+          message: e.message,
+        });
         return <manifold-toast alert-type="error">{e.message}</manifold-toast>; // show error to user
       }
     };